Adding api management connection details
