+++ Localize/lang/strings.json
@@ -10,6 +10,7 @@
   "+iMTt7": "Check to select this connection",
   "+lpuMl": "Required. The value to return if the expression is 'false'.",
   "+xiT4Q": "Pagination",
+  "//nQG4": "Select an api from apim instance",
   "/AtN16": "(UTC+09:30) Darwin",
   "/B2CTz": "(UTC+08:00) Beijing, Chongqing, Hong Kong, Urumqi",
   "/I/U8L": "Body",
@@ -54,7 +55,6 @@
   "11qu8f": "Enter a valid integer.",
   "14lYtE": "18",
   "1A1P5b": "Comment",
-  "1ANUHk": "Loading functions...",
   "1D047X": "Required. The value to convert to XML.",
   "1Fn5n+": "Required. The URI encoded string.",
   "1NBvKu": "Convert the parameter argument to a floating-point number",
@@ -378,6 +378,7 @@
   "LOwePI": "Invalid type on split on value '{splitOn}', split on not in array.",
   "LR/3Lr": "Configure",
   "LS8rfZ": "Returns the scheme from a URI",
+  "LV/BTE": "Loading Api Management service instances...",
   "LZm3ze": "Add a parallel branch",
   "Lft/is": "Add new",
   "Li6x/B": "Missing required properties '{0}' for authentication type '{1}'",
@@ -629,6 +630,7 @@
   "_+iMTt7.comment": "aria label description for check button",
   "_+lpuMl.comment": "Required value parameter to return given if function is false",
   "_+xiT4Q.comment": "title for pagination setting",
+  "_//nQG4.comment": "Label for api selection",
   "_/AtN16.comment": "Timezone value ",
   "_/B2CTz.comment": "Timezone value ",
   "_/I/U8L.comment": "Label text for request or response body",
@@ -673,7 +675,6 @@
   "_11qu8f.comment": "Error validation message",
   "_14lYtE.comment": "Hour of the day",
   "_1A1P5b.comment": "Comment Label",
-  "_1ANUHk.comment": "Message to show under the loading icon when loading functions",
   "_1D047X.comment": "Required string parameter to be converted using xml function",
   "_1Fn5n+.comment": "Required URI encoded string parameter to be converted using uriComponentToBinary function",
   "_1NBvKu.comment": "Label for description of custom float Function",
@@ -997,6 +998,7 @@
   "_LOwePI.comment": "Error message for when split on array is invalid.",
   "_LR/3Lr.comment": "Configure",
   "_LS8rfZ.comment": "Label for description of custom uriScheme Function",
+  "_LV/BTE.comment": "Text for loading apim service instances",
   "_LZm3ze.comment": "Text for button to add a parallel branch",
   "_Lft/is.comment": "Button to add a new connection",
   "_Li6x/B.comment": "Error message when missing multiple required authentication properties",
@@ -1376,7 +1378,6 @@
   "_iq+tiv.comment": "Hour of the day",
   "_iql+jn.comment": "This is a period in time in seconds. {seconds} is replaced by the number and s is an abbreviation of seconds",
   "_isnFBi.comment": "Message on switching levels with nodes/mappings not connected to a target schema node",
-  "_ittN0l.comment": "Message to show when no functions are found",
   "_iy8rNf.comment": "Button text for running test",
   "_j1VU25.comment": "Timezone value ",
   "_j2MU0b.comment": "Label for description of custom removeProperty Function",
@@ -1804,7 +1805,6 @@
   "iq+tiv": "11",
   "iql+jn": "{seconds}s",
   "isnFBi": "Elements and mappings not connected to a target element are removed.",
-  "ittN0l": "No resources of this type found under this subscription.",
   "iy8rNf": "Test",
   "j1VU25": "(UTC+14:00) Kiritimati Island",
   "j2MU0b": "Returns an object with a property removed",
