+++ libs/designer/src/lib/core/state/designerOptions/designerOptionsInterfaces.ts
@@ -7,6 +7,7 @@ import type {
   ISearchService,
   IOAuthService,
   IWorkflowService,
+  IApiManagementService,
 } from '@microsoft/designer-client-services-logic-apps';
 
 export interface DesignerOptionsState {
@@ -26,4 +27,5 @@ export interface ServiceOptions {
   loggerService?: ILoggerService;
   oAuthService: IOAuthService;
   workflowService: IWorkflowService;
+  apimService?: IApiManagementService;
 }
