+++ libs/designer/src/lib/core/utils/connectors/connections.ts
@@ -1,9 +1,10 @@
-import { ConnectionService } from '@microsoft/designer-client-services-logic-apps';
-import type { AssistedConnectionProps } from '@microsoft/designer-ui';
-import { getIntl } from '@microsoft/intl-logic-apps';
-import type { Connector } from '@microsoft/utils-logic-apps';
 import type { ConnectionReference } from '../../../common/models/workflow';
 import type { ConnectionsStoreState } from '../../state/connection/connectionSlice';
+import { ApiManagementService, ConnectionService } from '@microsoft/designer-client-services-logic-apps';
+import type { AssistedConnectionProps } from '@microsoft/designer-ui';
+import { getIntl } from '@microsoft/intl-logic-apps';
+import type { Connector, OperationManifest } from '@microsoft/utils-logic-apps';
+import { ConnectionType } from '@microsoft/utils-logic-apps';
 
 export function getConnectionId(state: ConnectionsStoreState, nodeId: string): string {
   const { connectionsMapping, connectionReferences } = state;
@@ -16,34 +17,33 @@ export function getConnectionReference(state: ConnectionsStoreState, nodeId: str
   return connectionReferences[connectionsMapping[nodeId] ?? ''];
 }
 
-export function getAssistedConnectionProps(connector: Connector): AssistedConnectionProps | undefined {
-  const isAzureFunctionConnection = connector.properties.capabilities?.includes('azureConnection') ?? false;
-  const isApiManagementConnection = connector.properties.capabilities?.includes('azureAPIManagement') ?? false;
+export function getAssistedConnectionProps(connector: Connector, manifest?: OperationManifest): AssistedConnectionProps | undefined {
+  const hasAzureConnection = connector.properties.capabilities?.includes('azureConnection') ?? false;
 
-  if (!isAzureFunctionConnection && !isApiManagementConnection) {
+  if (!hasAzureConnection && !manifest) {
     return undefined;
   }
 
   const intl = getIntl();
   const headers = [
-    intl.formatMessage({ defaultMessage: 'Name', description: 'Header for resource name', }),
-    intl.formatMessage({ defaultMessage: 'Resource Group', description: 'Header for resource group name', }),
-    intl.formatMessage({ defaultMessage: 'Location', description: 'Header for resource lcoation', }),
+    intl.formatMessage({ defaultMessage: 'Name', description: 'Header for resource name' }),
+    intl.formatMessage({ defaultMessage: 'Resource Group', description: 'Header for resource group name' }),
+    intl.formatMessage({ defaultMessage: 'Location', description: 'Header for resource lcoation' }),
   ];
-  if (isAzureFunctionConnection) {
+  if (manifest?.properties.connection?.type === ConnectionType.Function) {
     const functionAppsCallback = () => ConnectionService().fetchFunctionApps();
     const functionsCallback = (functionAppId?: string) => ConnectionService().fetchFunctionAppsFunctions(functionAppId ?? '');
     const functionAppsLoadingText = intl.formatMessage({
       defaultMessage: 'Loading Function Apps...',
       description: 'Text for loading function apps',
     });
-  
+
     const functionAppsLabel = intl.formatMessage({
       defaultMessage: 'Select a function app function',
       description: 'Label for function app selection',
     });
 
-    const getColumns = (functionApp: any) => ([functionApp?.name, functionApp?.properties?.resourceGroup, functionApp?.location]);
+    const getColumns = (functionApp: any) => [functionApp?.name, functionApp?.properties?.resourceGroup, functionApp?.location];
 
     return {
       resourceType: 'functionApps',
@@ -53,9 +53,67 @@ export function getAssistedConnectionProps(connector: Connector): AssistedConnec
       getColumns,
       getResourcesCallback: functionAppsCallback,
       resourcesLoadingText: functionAppsLoadingText,
-      fetchSubResourcesCallback: functionsCallback
+      getSubResourceName: (azureFunction: any) => azureFunction.name.split('/')[1],
+      fetchSubResourcesCallback: functionsCallback,
     };
-  } else {
-    return undefined;
+  } else if (manifest?.properties.connection?.type === ConnectionType.ApiManagement) {
+    const apiInstancesCallback = () => ApiManagementService().fetchApiManagementInstances();
+    const apisCallback = (apimId?: string) => ApiManagementService().fetchApisInApiM(apimId ?? '');
+    const apimInstancesLoadingText = intl.formatMessage({
+      defaultMessage: 'Loading Api Management service instances...',
+      description: 'Text for loading apim service instances',
+    });
+
+    const apisLabel = intl.formatMessage({
+      defaultMessage: 'Select an api from apim instance',
+      description: 'Label for api selection',
+    });
+
+    const getColumns = (apimInstance: any) => [apimInstance?.name, apimInstance?.id.split('/')[4], apimInstance?.location];
+
+    return {
+      resourceType: 'apimInstances',
+      subResourceType: 'apimApis',
+      title: apisLabel,
+      headers,
+      getColumns,
+      getResourcesCallback: apiInstancesCallback,
+      resourcesLoadingText: apimInstancesLoadingText,
+      getSubResourceName: (api: any) => api.name,
+      fetchSubResourcesCallback: apisCallback,
+    };
   }
+
+  return undefined;
+}
+
+export async function getConnectionParametersForAzureConnection(connectionType?: ConnectionType, selectedSubResource?: any): Promise<any> {
+  if (connectionType === ConnectionType.Function) {
+    const functionId = selectedSubResource?.id;
+    const authCodeValue = await ConnectionService().fetchFunctionKey(functionId);
+    const triggerUrl = selectedSubResource?.properties?.invoke_url_template;
+    return {
+      function: { id: functionId },
+      triggerUrl,
+      authentication: {
+        type: 'QueryString',
+        name: 'Code',
+        value: authCodeValue,
+      },
+    };
+  } else if (connectionType === ConnectionType.ApiManagement) {
+    // TODO - Need to find apps which have authentication set, check with Alex.
+    const apimApiId = selectedSubResource?.id;
+    const apiSwagger = await ApiManagementService().fetchApiMSwagger(apimApiId);
+    const baseUrl = apiSwagger.host ?? 'NotFound';
+    const subscriptionKey = apiSwagger.securityDefinitions?.apiKeyHeader?.name ?? 'NotFound';
+
+    return {
+      apiId: apimApiId,
+      baseUrl,
+      subscriptionKey,
+    };
+  }
+
+  return {};
 }
