+++ libs/services/designer-client-services/src/index.ts
@@ -13,3 +13,4 @@ export * from './lib/logging/logEntry';
 export * from './lib/base';
 export * from './lib/standard';
 export * from './lib/consumption';
+export * from './lib/apimanagement';
