+++ libs/services/designer-client-services/src/lib/base/connection.ts
@@ -1,5 +1,4 @@
 /* eslint-disable no-param-reassign */
-import { azureFunctionConnectorId } from '../base/operationmanifest';
 import type { HttpResponse } from '../common/exceptions/service';
 import type {
   ConnectionCreationInfo,
@@ -9,59 +8,16 @@ import type {
   ConnectorWithSwagger,
 } from '../connection';
 import type { HttpRequestOptions, IHttpClient, QueryParameters } from '../httpClient';
-import type { Connection, ConnectionParameter, Connector, ManagedIdentity } from '@microsoft/utils-logic-apps';
+import type { Connection, Connector } from '@microsoft/utils-logic-apps';
 import {
   getUniqueName,
   HTTP_METHODS,
   UserErrorCode,
   UserException,
   isArmResourceId,
-  ConnectionParameterSource,
-  ConnectionType,
-  safeSetObjectPropertyValue,
   ArgumentException,
   equals,
 } from '@microsoft/utils-logic-apps';
-
-interface ServiceProviderConnectionModel {
-  parameterValues: Record<string, any>;
-  serviceProvider: {
-    id: string;
-  };
-  displayName?: string;
-}
-
-interface FunctionsConnectionModel {
-  function: {
-    id: string;
-  };
-  triggerUrl: string;
-  authentication: {
-    type: string;
-    name: string;
-    value: string;
-  };
-  displayName?: string;
-}
-
-interface APIManagementConnectionModel {
-  apiId: string;
-  baseUrl: string;
-  subscriptionKey: string;
-  authentication?: {
-    type: string;
-    name: string;
-    value: string;
-  };
-  displayName?: string;
-}
-
-interface ConnectionAndAppSetting<T> {
-  connectionKey: string;
-  connectionData: T;
-  settings: Record<string, string>;
-  pathLocation: string[];
-}
 
 export interface IApiHubServiceDetails {
   apiVersion: string;
@@ -77,21 +33,8 @@ export interface BaseConnectionServiceOptions {
   locale?: string;
   filterByLocation?: boolean;
   tenantId?: string;
-  workflowAppDetails?: {
-    appName: string;
-    identity?: ManagedIdentity;
-  };
-  readConnections: ReadConnectionsFunc;
-  writeConnection?: WriteConnectionFunc;
-  apiHubServiceDetails: IApiHubServiceDetails;
   httpClient: IHttpClient;
-  connectionCreationClients?: Record<string, ConnectionCreationClient>;
-}
-
-type CreateConnectionFunc = (connectionInfo: ConnectionCreationInfo, connectionName: string) => Promise<ConnectionCreationInfo>;
-
-interface ConnectionCreationClient {
-  connectionCreationFunc: CreateConnectionFunc;
+  apiHubServiceDetails: IApiHubServiceDetails;
 }
 
 export type getAccessTokenType = () => Promise<string>;
@@ -106,19 +49,6 @@ export interface LogicAppConsentResponse {
   value: ConsentLink[];
 }
 
-interface ConnectionsData {
-  managedApiConnections?: any;
-  serviceProviderConnections?: Record<string, ServiceProviderConnectionModel>;
-  functionConnections?: Record<string, FunctionsConnectionModel>;
-}
-
-type LocalConnectionModel = FunctionsConnectionModel | ServiceProviderConnectionModel | APIManagementConnectionModel;
-type ReadConnectionsFunc = () => Promise<ConnectionsData>;
-type WriteConnectionFunc = (connectionData: ConnectionAndAppSetting<LocalConnectionModel>) => Promise<void>;
-
-const serviceProviderLocation = 'serviceProviderConnections';
-const functionsLocation = 'functionConnections';
-
 export abstract class BaseConnectionService implements IConnectionService {
   protected _connections: Record<string, Connection> = {};
   protected _subscriptionResourceGroupWebUrl = '';
@@ -127,13 +57,11 @@ export abstract class BaseConnectionService implements IConnectionService {
   protected _vVersion: 'V1' | 'V2' = 'V1';
 
   constructor(public readonly options: BaseConnectionServiceOptions) {
-    const { apiHubServiceDetails, apiVersion, baseUrl, readConnections } = options;
+    const { apiVersion, baseUrl, apiHubServiceDetails } = options;
     if (!baseUrl) {
       throw new ArgumentException('baseUrl required');
     } else if (!apiVersion) {
       throw new ArgumentException('apiVersion required');
-    } else if (!readConnections) {
-      throw new ArgumentException('readConnections required');
     } else if (!apiHubServiceDetails) {
       throw new ArgumentException('apiHubServiceDetails required for workflow app');
     }
@@ -202,29 +130,8 @@ export abstract class BaseConnectionService implements IConnectionService {
     return connection;
   }
 
-  async getConnections(connectorId?: string): Promise<Connection[]> {
-    if (connectorId) {
-      return this.getConnectionsForConnector(connectorId);
-    }
-
-    const [localConnections, apiHubConnections] = await Promise.all([this.options.readConnections(), this.getConnectionsInApiHub()]);
-    const serviceProviderConnections = (localConnections[serviceProviderLocation] || {}) as Record<string, ServiceProviderConnectionModel>;
-    const functionConnections = (localConnections[functionsLocation] || {}) as Record<string, FunctionsConnectionModel>;
-
-    this._allConnectionsInitialized = true;
-    return [
-      ...Object.keys(serviceProviderConnections).map((key) => {
-        const connection = convertServiceProviderConnectionDataToConnection(key, serviceProviderConnections[key]);
-        this._connections[connection.id] = connection;
-        return connection;
-      }),
-      ...Object.keys(functionConnections).map((key) => {
-        const connection = convertFunctionsConnectionDataToConnection(key, functionConnections[key]);
-        this._connections[connection.id] = connection;
-        return connection;
-      }),
-      ...apiHubConnections,
-    ];
+  async getConnections(_connectorId?: string): Promise<Connection[]> {
+    throw new Error('Should be implemented in extending class');
   }
 
   abstract createConnection(
@@ -375,39 +282,6 @@ export abstract class BaseConnectionService implements IConnectionService {
 
   protected tryParseErrorMessage(error: any, defaultErrorMessage?: string): string {
     return error?.message ?? error?.Message ?? error?.error?.message ?? error?.responseText ?? defaultErrorMessage ?? 'Unknown error';
-  }
-
-  protected getConnectionsConfiguration(
-    connectionName: string,
-    connectionInfo: ConnectionCreationInfo,
-    connectorId: string,
-    parametersMetadata: ConnectionParametersMetadata
-  ): {
-    connectionsData: ConnectionAndAppSetting<LocalConnectionModel>;
-    connection: Connection;
-  } {
-    const connectionType = parametersMetadata?.connectionMetadata?.type;
-    let connectionsData;
-    let connection;
-    switch (connectionType) {
-      case ConnectionType.Function: {
-        connectionsData = convertToFunctionsConnectionsData(connectionName, connectionInfo);
-        connection = convertFunctionsConnectionDataToConnection(
-          connectionsData.connectionKey,
-          connectionsData.connectionData as FunctionsConnectionModel
-        );
-        break;
-      }
-      default: {
-        connectionsData = convertToServiceProviderConnectionsData(connectionName, connectorId, connectionInfo, parametersMetadata);
-        connection = convertServiceProviderConnectionDataToConnection(
-          connectionsData.connectionKey,
-          connectionsData.connectionData as ServiceProviderConnectionModel
-        );
-        break;
-      }
-    }
-    return { connectionsData, connection };
   }
 
   protected async getConnectionsForConnector(connectorId: string): Promise<Connection[]> {
@@ -567,131 +441,3 @@ export abstract class BaseConnectionService implements IConnectionService {
 type ConnectionsResponse = {
   value: Connection[];
 };
-
-function convertServiceProviderConnectionDataToConnection(
-  connectionKey: string,
-  connectionData: ServiceProviderConnectionModel
-): Connection {
-  const {
-    displayName,
-    serviceProvider: { id: apiId },
-  } = connectionData;
-
-  return {
-    name: connectionKey,
-    id: `${apiId}/connections/${connectionKey}`,
-    type: 'connections',
-    properties: {
-      api: { id: apiId } as any,
-      createdTime: '',
-      connectionParameters: {},
-      displayName: displayName as string,
-      statuses: [{ status: 'Connected' }],
-      overallStatus: 'Connected',
-      testLinks: [],
-    },
-  };
-}
-
-function convertToServiceProviderConnectionsData(
-  connectionKey: string,
-  connectorId: string,
-  connectionInfo: ConnectionCreationInfo,
-  connectionParameterMetadata: ConnectionParametersMetadata
-): ConnectionAndAppSetting<ServiceProviderConnectionModel> {
-  const {
-    displayName,
-    connectionParameters: connectionParameterValues,
-    connectionParametersSet: connectionParametersSetValues,
-  } = connectionInfo;
-  const connectionParameters = connectionParametersSetValues
-    ? connectionParameterMetadata.connectionParameterSet?.parameters
-    : (connectionParameterMetadata.connectionParameters as Record<string, ConnectionParameter>);
-  const parameterValues = connectionParametersSetValues
-    ? Object.keys(connectionParametersSetValues.values).reduce(
-        (result: Record<string, any>, currentKey: string) => ({
-          ...result,
-          [currentKey]: connectionParametersSetValues.values[currentKey].value,
-        }),
-        {}
-      )
-    : (connectionParameterValues as Record<string, any>);
-
-  const connectionsData: ConnectionAndAppSetting<ServiceProviderConnectionModel> = {
-    connectionKey,
-    connectionData: {
-      parameterValues: {},
-      serviceProvider: { id: connectorId },
-      displayName,
-    },
-    settings: {},
-    pathLocation: [serviceProviderLocation],
-  };
-
-  for (const parameterKey of Object.keys(parameterValues)) {
-    const connectionParameter = connectionParameters?.[parameterKey] as ConnectionParameter;
-    let parameterValue = parameterValues[parameterKey];
-    if (connectionParameter?.parameterSource === ConnectionParameterSource.AppConfiguration) {
-      const appSettingName = `${escapeSpecialChars(connectionKey)}_${escapeSpecialChars(parameterKey)}`;
-      connectionsData.settings[appSettingName] = parameterValues[parameterKey];
-
-      parameterValue = `@appsetting('${appSettingName}')`;
-    }
-
-    safeSetObjectPropertyValue(
-      connectionsData.connectionData.parameterValues,
-      [...(connectionParameter?.uiDefinition?.constraints?.propertyPath ?? []), parameterKey],
-      parameterValue
-    );
-  }
-
-  return connectionsData;
-}
-
-function convertToFunctionsConnectionsData(
-  connectionKey: string,
-  connectionInfo: ConnectionCreationInfo
-): ConnectionAndAppSetting<FunctionsConnectionModel> {
-  const { displayName, connectionParameters } = connectionInfo;
-  const authentication = connectionParameters?.['authentication'];
-  const functionAppKey = authentication.value;
-  const appSettingName = `${escapeSpecialChars(connectionKey)}_functionAppKey`;
-
-  authentication.value = `@appsetting('${appSettingName}')`;
-
-  return {
-    connectionKey,
-    connectionData: {
-      function: connectionParameters?.['function'],
-      triggerUrl: connectionParameters?.['triggerUrl'],
-      authentication,
-      displayName,
-    },
-    settings: { [appSettingName]: functionAppKey },
-    pathLocation: [functionsLocation],
-  };
-}
-
-function convertFunctionsConnectionDataToConnection(connectionKey: string, connectionData: FunctionsConnectionModel): Connection {
-  const { displayName } = connectionData;
-
-  return {
-    name: connectionKey,
-    id: `${azureFunctionConnectorId}/connections/${connectionKey}`,
-    type: 'connections',
-    properties: {
-      api: { id: azureFunctionConnectorId } as any,
-      createdTime: '',
-      connectionParameters: {},
-      displayName: displayName as string,
-      overallStatus: 'Connected',
-      statuses: [{ status: 'Connected' }],
-      testLinks: [],
-    },
-  };
-}
-
-function escapeSpecialChars(value: string): string {
-  const escapedUnderscore = value.replace(/_/g, '__');
-  return escapedUnderscore.replace(/-/g, '_1');
-}
