+++ libs/services/designer-client-services/src/lib/standard/connection.ts
@@ -1,18 +1,24 @@
+import type { BaseConnectionServiceOptions } from '../base';
 import { BaseConnectionService } from '../base';
+import { apiManagementConnectorId, azureFunctionConnectorId } from '../base/operationmanifest';
 import type { ConnectionCreationInfo, ConnectionParametersMetadata, CreateConnectionResult } from '../connection';
 import { LoggerService } from '../logger';
 import { LogEntryLevel } from '../logging/logEntry';
 import type { IOAuthPopup } from '../oAuth';
 import { OAuthService } from '../oAuth';
 import { getIntl } from '@microsoft/intl-logic-apps';
-import type { Connection, Connector, ManagedIdentity } from '@microsoft/utils-logic-apps';
+import type { Connection, ConnectionParameter, Connector, ManagedIdentity } from '@microsoft/utils-logic-apps';
 import {
+  ArgumentException,
   AssertionErrorCode,
   AssertionException,
+  ConnectionParameterSource,
+  ConnectionType,
   ResourceIdentityType,
   equals,
   isArmResourceId,
   isIdentityAssociatedWithLogicApp,
+  safeSetObjectPropertyValue,
 } from '@microsoft/utils-logic-apps';
 
 interface ConnectionAcl {
@@ -31,10 +37,117 @@ interface ConnectionAcl {
   type: string;
 }
 
+interface ServiceProviderConnectionModel {
+  parameterValues: Record<string, any>;
+  serviceProvider: {
+    id: string;
+  };
+  displayName?: string;
+}
+
+interface FunctionsConnectionModel {
+  function: {
+    id: string;
+  };
+  triggerUrl: string;
+  authentication: {
+    type: string;
+    name: string;
+    value: string;
+  };
+  displayName?: string;
+}
+
+interface APIManagementConnectionModel {
+  apiId: string;
+  baseUrl: string;
+  subscriptionKey: string;
+  authentication?: {
+    type: string;
+    name: string;
+    value: string;
+  };
+  displayName?: string;
+}
+
+interface ConnectionAndAppSetting<T> {
+  connectionKey: string;
+  connectionData: T;
+  settings: Record<string, string>;
+  pathLocation: string[];
+}
+
+interface ConnectionsData {
+  managedApiConnections?: any;
+  serviceProviderConnections?: Record<string, ServiceProviderConnectionModel>;
+  functionConnections?: Record<string, FunctionsConnectionModel>;
+  apiManagementConnections?: Record<string, APIManagementConnectionModel>;
+}
+
+type LocalConnectionModel = FunctionsConnectionModel | ServiceProviderConnectionModel | APIManagementConnectionModel;
+type ReadConnectionsFunc = () => Promise<ConnectionsData>;
+type WriteConnectionFunc = (connectionData: ConnectionAndAppSetting<LocalConnectionModel>) => Promise<void>;
+
+const serviceProviderLocation = 'serviceProviderConnections';
+const functionsLocation = 'functionConnections';
+const apimLocation = 'apiManagementConnections';
+
+export interface StandardConnectionServiceOptions extends BaseConnectionServiceOptions {
+  workflowAppDetails?: {
+    appName: string;
+    identity?: ManagedIdentity;
+  };
+  readConnections: ReadConnectionsFunc;
+  writeConnection?: WriteConnectionFunc;
+  connectionCreationClients?: Record<string, ConnectionCreationClient>;
+}
+
+type CreateConnectionFunc = (connectionInfo: ConnectionCreationInfo, connectionName: string) => Promise<ConnectionCreationInfo>;
+
+interface ConnectionCreationClient {
+  connectionCreationFunc: CreateConnectionFunc;
+}
+
 export class StandardConnectionService extends BaseConnectionService {
-  constructor(options: any) {
+  constructor(public override readonly options: StandardConnectionServiceOptions) {
     super(options);
+    const { readConnections } = options;
+    if (!readConnections) {
+      throw new ArgumentException('readConnections required');
+    }
+
     this._vVersion = 'V2';
+  }
+
+  override async getConnections(connectorId?: string): Promise<Connection[]> {
+    if (connectorId) {
+      return this.getConnectionsForConnector(connectorId);
+    }
+
+    const [localConnections, apiHubConnections] = await Promise.all([this.options.readConnections(), this.getConnectionsInApiHub()]);
+    const serviceProviderConnections = (localConnections[serviceProviderLocation] || {}) as Record<string, ServiceProviderConnectionModel>;
+    const functionConnections = (localConnections[functionsLocation] || {}) as Record<string, FunctionsConnectionModel>;
+    const apimConnections = (localConnections[apimLocation] || {}) as Record<string, APIManagementConnectionModel>;
+
+    this._allConnectionsInitialized = true;
+    return [
+      ...Object.keys(serviceProviderConnections).map((key) => {
+        const connection = convertServiceProviderConnectionDataToConnection(key, serviceProviderConnections[key]);
+        this._connections[connection.id] = connection;
+        return connection;
+      }),
+      ...Object.keys(functionConnections).map((key) => {
+        const connection = convertFunctionsConnectionDataToConnection(key, functionConnections[key]);
+        this._connections[connection.id] = connection;
+        return connection;
+      }),
+      ...Object.keys(apimConnections).map((key) => {
+        const connection = convertApimConnectionDataToConnection(key, apimConnections[key]);
+        this._connections[connection.id] = connection;
+        return connection;
+      }),
+      ...apiHubConnections,
+    ];
   }
 
   async createConnection(
@@ -104,7 +217,7 @@ export class StandardConnectionService extends BaseConnectionService {
       throw new AssertionException(AssertionErrorCode.CALLBACK_NOTREGISTERED, 'Callback for write connection is not passed in service.');
     }
 
-    const { connectionsData, connection } = this.getConnectionsConfiguration(
+    const { connectionsData, connection } = this._getConnectionsConfiguration(
       connectionName,
       connectionInfo,
       connectorId,
@@ -302,4 +415,217 @@ export class StandardConnectionService extends BaseConnectionService {
       return { errorMessage: this.tryParseErrorMessage(error) };
     }
   }
+
+  private _getConnectionsConfiguration(
+    connectionName: string,
+    connectionInfo: ConnectionCreationInfo,
+    connectorId: string,
+    parametersMetadata: ConnectionParametersMetadata
+  ): {
+    connectionsData: ConnectionAndAppSetting<LocalConnectionModel>;
+    connection: Connection;
+  } {
+    const connectionType = parametersMetadata?.connectionMetadata?.type;
+    let connectionsData;
+    let connection;
+    switch (connectionType) {
+      case ConnectionType.Function: {
+        connectionsData = convertToFunctionsConnectionsData(connectionName, connectionInfo);
+        connection = convertFunctionsConnectionDataToConnection(
+          connectionsData.connectionKey,
+          connectionsData.connectionData as FunctionsConnectionModel
+        );
+        break;
+      }
+      case ConnectionType.ApiManagement: {
+        connectionsData = convertToApimConnectionsData(connectionName, connectionInfo);
+        connection = convertApimConnectionDataToConnection(
+          connectionsData.connectionKey,
+          connectionsData.connectionData as APIManagementConnectionModel
+        );
+        break;
+      }
+      default: {
+        connectionsData = convertToServiceProviderConnectionsData(connectionName, connectorId, connectionInfo, parametersMetadata);
+        connection = convertServiceProviderConnectionDataToConnection(
+          connectionsData.connectionKey,
+          connectionsData.connectionData as ServiceProviderConnectionModel
+        );
+        break;
+      }
+    }
+    return { connectionsData, connection };
+  }
+}
+
+function convertServiceProviderConnectionDataToConnection(
+  connectionKey: string,
+  connectionData: ServiceProviderConnectionModel
+): Connection {
+  const {
+    displayName,
+    serviceProvider: { id: apiId },
+  } = connectionData;
+
+  return {
+    name: connectionKey,
+    id: `${apiId}/connections/${connectionKey}`,
+    type: 'connections',
+    properties: {
+      api: { id: apiId } as any,
+      createdTime: '',
+      connectionParameters: {},
+      displayName: displayName as string,
+      statuses: [{ status: 'Connected' }],
+      overallStatus: 'Connected',
+      testLinks: [],
+    },
+  };
+}
+
+function convertApimConnectionDataToConnection(connectionKey: string, connectionData: APIManagementConnectionModel): Connection {
+  const { displayName } = connectionData;
+
+  return {
+    name: connectionKey,
+    id: `${apiManagementConnectorId}/connections/${connectionKey}`,
+    type: 'connections',
+    properties: {
+      api: { id: apiManagementConnectorId } as any,
+      createdTime: '',
+      connectionParameters: {},
+      displayName: displayName as string,
+      overallStatus: 'Connected',
+      statuses: [{ status: 'Connected' }],
+      testLinks: [],
+    },
+  };
+}
+
+function convertFunctionsConnectionDataToConnection(connectionKey: string, connectionData: FunctionsConnectionModel): Connection {
+  const { displayName } = connectionData;
+
+  return {
+    name: connectionKey,
+    id: `${azureFunctionConnectorId}/connections/${connectionKey}`,
+    type: 'connections',
+    properties: {
+      api: { id: azureFunctionConnectorId } as any,
+      createdTime: '',
+      connectionParameters: {},
+      displayName: displayName as string,
+      overallStatus: 'Connected',
+      statuses: [{ status: 'Connected' }],
+      testLinks: [],
+    },
+  };
+}
+
+function convertToServiceProviderConnectionsData(
+  connectionKey: string,
+  connectorId: string,
+  connectionInfo: ConnectionCreationInfo,
+  connectionParameterMetadata: ConnectionParametersMetadata
+): ConnectionAndAppSetting<ServiceProviderConnectionModel> {
+  const {
+    displayName,
+    connectionParameters: connectionParameterValues,
+    connectionParametersSet: connectionParametersSetValues,
+  } = connectionInfo;
+  const connectionParameters = connectionParametersSetValues
+    ? connectionParameterMetadata.connectionParameterSet?.parameters
+    : (connectionParameterMetadata.connectionParameters as Record<string, ConnectionParameter>);
+  const parameterValues = connectionParametersSetValues
+    ? Object.keys(connectionParametersSetValues.values).reduce(
+        (result: Record<string, any>, currentKey: string) => ({
+          ...result,
+          [currentKey]: connectionParametersSetValues.values[currentKey].value,
+        }),
+        {}
+      )
+    : (connectionParameterValues as Record<string, any>);
+
+  const connectionsData: ConnectionAndAppSetting<ServiceProviderConnectionModel> = {
+    connectionKey,
+    connectionData: {
+      parameterValues: {},
+      serviceProvider: { id: connectorId },
+      displayName,
+    },
+    settings: {},
+    pathLocation: [serviceProviderLocation],
+  };
+
+  for (const parameterKey of Object.keys(parameterValues)) {
+    const connectionParameter = connectionParameters?.[parameterKey] as ConnectionParameter;
+    let parameterValue = parameterValues[parameterKey];
+    if (connectionParameter?.parameterSource === ConnectionParameterSource.AppConfiguration) {
+      const appSettingName = `${escapeSpecialChars(connectionKey)}_${escapeSpecialChars(parameterKey)}`;
+      connectionsData.settings[appSettingName] = parameterValues[parameterKey];
+
+      parameterValue = `@appsetting('${appSettingName}')`;
+    }
+
+    safeSetObjectPropertyValue(
+      connectionsData.connectionData.parameterValues,
+      [...(connectionParameter?.uiDefinition?.constraints?.propertyPath ?? []), parameterKey],
+      parameterValue
+    );
+  }
+
+  return connectionsData;
+}
+
+function convertToFunctionsConnectionsData(
+  connectionKey: string,
+  connectionInfo: ConnectionCreationInfo
+): ConnectionAndAppSetting<FunctionsConnectionModel> {
+  const { displayName, connectionParameters } = connectionInfo;
+  const authentication = connectionParameters?.['authentication'];
+  const functionAppKey = authentication.value;
+  const appSettingName = `${escapeSpecialChars(connectionKey)}_functionAppKey`;
+
+  authentication.value = `@appsetting('${appSettingName}')`;
+
+  return {
+    connectionKey,
+    connectionData: {
+      function: connectionParameters?.['function'],
+      triggerUrl: connectionParameters?.['triggerUrl'],
+      authentication,
+      displayName,
+    },
+    settings: { [appSettingName]: functionAppKey },
+    pathLocation: [functionsLocation],
+  };
+}
+
+function convertToApimConnectionsData(
+  connectionKey: string,
+  connectionInfo: ConnectionCreationInfo
+): ConnectionAndAppSetting<APIManagementConnectionModel> {
+  const { displayName, connectionParameters } = connectionInfo;
+  const subscriptionKey = connectionParameters?.['subscripitonKey'];
+  const functionAppKey = subscriptionKey.value;
+  const appSettingName = `${escapeSpecialChars(connectionKey)}_SubscriptionKey`;
+
+  subscriptionKey.value = `@appsetting('${appSettingName}')`;
+
+  return {
+    connectionKey,
+    connectionData: {
+      apiId: connectionParameters?.['apiId'],
+      baseUrl: connectionParameters?.['baseUrl'],
+      subscriptionKey,
+      authentication: connectionParameters?.['authentication'],
+      displayName,
+    },
+    settings: { [appSettingName]: functionAppKey },
+    pathLocation: [apimLocation],
+  };
+}
+
+function escapeSpecialChars(value: string): string {
+  const escapedUnderscore = value.replace(/_/g, '__');
+  return escapedUnderscore.replace(/-/g, '_1');
 }
