+++ libs/services/intl/src/compiled-lang/strings.en-XA.json
@@ -65,6 +65,12 @@
       "value": "Ṕâĝíńâṭíŏń"
     }
   ],
+  "//nQG4": [
+    {
+      "type": 0,
+      "value": "Ṡèĺèćṭ âń âṗí ḟŕŏṁ âṗíṁ íńśṭâńćè"
+    }
+  ],
   "/AtN16": [
     {
       "type": 0,
@@ -331,12 +337,6 @@
     {
       "type": 0,
       "value": "Ḉŏṁṁèńṭ"
-    }
-  ],
-  "1ANUHk": [
-    {
-      "type": 0,
-      "value": "Ļŏâḋíńĝ ḟůńćṭíŏńś..."
     }
   ],
   "1D047X": [
@@ -2495,6 +2495,12 @@
     {
       "type": 0,
       "value": "Ŕèṭůŕńś ṭḫè śćḫèṁè ḟŕŏṁ â ŨŔḬ"
+    }
+  ],
+  "LV/BTE": [
+    {
+      "type": 0,
+      "value": "Ļŏâḋíńĝ Ḁṗí Ḿâńâĝèṁèńṭ śèŕṿíćè íńśṭâńćèś..."
     }
   ],
   "LZm3ze": [
@@ -4981,12 +4987,6 @@
     {
       "type": 0,
       "value": "Ḕĺèṁèńṭś âńḋ ṁâṗṗíńĝś ńŏṭ ćŏńńèćṭèḋ ṭŏ â ṭâŕĝèṭ èĺèṁèńṭ âŕè ŕèṁŏṿèḋ."
-    }
-  ],
-  "ittN0l": [
-    {
-      "type": 0,
-      "value": "Ŋŏ ŕèśŏůŕćèś ŏḟ ṭḫíś ṭẏṗè ḟŏůńḋ ůńḋèŕ ṭḫíś śůḃśćŕíṗṭíŏń."
     }
   ],
   "iy8rNf": [
